Updated @angular/cli from 1.7.4 to 8.3.5
